# apps/api/openapi/eiah-public.yaml
openapi: "3.1.0"
info:
  title: EIAH Public API
  version: 0.1.0
...
servers:
  - url: https://api.eiah.local
paths:
  /runs:
    get:
      operationId: listRuns
      summary: Lista runs com filtros e paginação
      parameters:
        - in: query
          name: agent
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum:
              - success
              - error
              - blocked
        - in: query
          name: projectId
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Run'
                  total:
                    type: integer
  /runs/{id}:
    get:
      operationId: getRun
      summary: Detalhe de um run
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
...
components:
  schemas:
    Run:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        agent:
          type: string
        status:
          type: string
          enum:
            - success
            - error
            - blocked
        request:
          type: object
        response:
          type: object
        costCents:
          type: integer
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        meta:
          type: object
          properties:
            traceId:
              type: string
            tookMs:
              type: number
...